Fix handling of Quantity sub-types Is STU3 the Quantity sub-types Age, Distance, Duration, Count and Money are actual FHIR data types, whereas in DSTU2 they are simple constraints on the Quantity data type. This commit split the corresponding classes in STU3 and DSTU2 specific ones, handling them appropriately in the two cases: serialized as Quantity in DSTU2, serialized as their own type in STU3.
Note that SimpleQuantity is a constraint in both DSTU2 and STU3, so it is still a shared class.

Also: generate the constraints properties also for elements (data types) and not just for resources. They are still not actually used for element validations, but they at least act as documentation (this replaces comments describing the constraints for sub-classes of Quantity)

Also: the code generation script creates the sub-directories for version-specific code files if they do not already exist (previously it crashed if those directories did not exist)