remove the `'` from custom date/time format strings
TryParseExact uses a stringbulider for each delimiter literal when it is surrounded by single quotes, which can end up with a lot of strings on the heap.

Since we are specifying the invariant culture for each call, the `'` are not needed and we can just use the unescaped delimiter characters.